fix(generator): conditionally add --go-grpc_out flag
When a BUILD.bazel file uses the 'go_grpc_library' rule, the
protoc-gen-go-grpc plugin must be invoked. This is done by adding
the --go-grpc_out flag to the protoc command.

This commit updates the Bazel parser to detect the presence of the
go_grpc_library rule and updates the protoc command builder to
conditionally add the --go-grpc_out flag when it is present.

This fixes an issue where gRPC service code was not being generated
for some APIs.

Update integration test to use chronicle/v1.