fix(container): Use public base image for development
This change temporarily replaces the MOSS-compliant base image
() with a public equivalent
().

This is a pragmatic fix to unblock local development for users who may
not have the necessary IAM permissions to pull the restricted,
Google-internal base image.

A  comment has been added to the Dockerfile to ensure that it is
switched back to the MOSS-compliant image for official production
builds.