feat(postprocessor): Add go mod tidy step
This change adds a  step to the post-processor, which
aligns it with the legacy implementation.

The  command is run for both new and existing modules to
ensure that the module's dependencies are always correct and
up-to-date with any changes in the generated  files.

Comprehensive unit tests have been added to verify that
is called in all success scenarios and that errors from the command are
correctly propagated.